# app/svc/docs/swagger.yaml
basePath: /
definitions:
...
  handler.AdminProject:
    properties:
      createdAt:
        type: integer
      createdBy:
        type: string
      id:
        type: string
      memberCount:
        type: integer
      name:
        type: string
      roleLabels:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  handler.Annotation:
    properties:
      authorUserId:
        type: string
      createdAt:
        type: integer
      entryId:
        type: string
      id:
        type: string
      sharedWith:
        type: string
      text:
        type: string
      updatedAt:
        type: integer
      visibility:
        type: string
    type: object
  handler.AuthForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  handler.AuthLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  handler.AuthResetPasswordRequest:
    properties:
      newPassword:
        type: string
      token:
        type: string
    type: object
...
  handler.CreateAnnotationRequest:
    properties:
      entryId:
        type: string
      sharedWith:
        type: string
      text:
        type: string
      visibility:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.MeResponse:
    properties:
      active:
        type: boolean
      email:
        type: string
      role:
        type: string
      sub:
        type: string
    type: object
...
  handler.OkResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  handler.ProjectMember:
    properties:
      id:
        type: string
      joinedAt:
        type: integer
      projectId:
        type: string
      role:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: 'Backend-for-frontend: session management, campaigns, entries, annotations.'
  title: Quillit BFF API
  version: "1.0"
paths:
  /api/admin/projects:
    get:
      description: Returns all projects with member counts. Supports ?q= name search.
      parameters:
      - description: Search by project name
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.AdminProject'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List all projects (admin)
      tags:
      - admin
  /api/admin/projects/{id}/members:
    get:
      description: Returns all members for the given project, regardless of admin's
        own membership.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.ProjectMember'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List members of any project (admin)
      tags:
      - admin
  /api/admin/users:
    get:
      description: Proxies to auth-svc. Supports ?q= search and ?active= filter.
      parameters:
      - description: Search by email or username
        in: query
        name: q
        type: string
      - description: 'Filter: true or false'
        in: query
        name: active
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List users (admin)
      tags:
      - admin
  /api/admin/users/{id}:
    delete:
      description: Proxies to auth-svc. Permanently removes the user account.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete user (admin)
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: 'Proxies to auth-svc. Supports { active: bool } to enable/disable
        an account.'
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: '{ active: bool }'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update user (admin)
      tags:
      - admin
  /api/annotations:
    get:
      parameters:
      - description: Filter by entry ID
        in: query
        name: entryId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Annotation'
            type: array
      summary: List annotations
      tags:
      - annotations
    post:
      consumes:
      - application/json
      parameters:
      - description: Annotation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAnnotationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Annotation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create annotation
      tags:
      - annotations
  /api/annotations/{id}:
    delete:
      parameters:
      - description: Annotation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
      summary: Delete annotation
      tags:
      - annotations
    patch:
      consumes:
      - application/json
      description: Partial update — send only the fields to change.
      parameters:
      - description: Annotation ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAnnotationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Annotation'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update annotation
      tags:
      - annotations
  /api/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Proxies to auth-svc. Always returns 200 regardless of whether the
        email is registered.
      parameters:
      - description: Email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AuthForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
      summary: Forgot password
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Forwards credentials to auth-svc, creates an HTTP-only session
        cookie on success.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /api/auth/me:
    get:
      description: Returns identity decoded from the session JWT cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Current user
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Forwards credentials to auth-svc, creates an HTTP-only session
        cookie on success.
      parameters:
      - description: Registration credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AuthRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register
      tags:
      - auth
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Proxies to auth-svc.
      parameters:
      - description: Token and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AuthResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Reset password
      tags:
      - auth
...
  /api/auth/users/available:
    get:
      description: Proxies to auth-svc to check whether a username is free to register.
      parameters:
      - description: Username to check
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UsernameAvailableResponse'
      summary: Username availability
      tags:
      - auth
  /api/campaigns:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Campaign'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List campaigns
      tags:
      - campaigns
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Campaign'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create campaign
      tags:
      - campaigns
  /api/campaigns/{campaignId}/players:
    post:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: campaignId
        required: true
        type: string
      - description: Player name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AddPlayerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Player'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Add player to campaign
      tags:
      - campaigns
  /api/campaigns/{campaignId}/players/{playerId}:
    delete:
      parameters:
      - description: Campaign ID
        in: path
        name: campaignId
        required: true
        type: string
      - description: Player ID
        in: path
        name: playerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
      summary: Remove player from campaign
      tags:
      - campaigns
  /api/campaigns/{id}:
    delete:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
      summary: Delete campaign
      tags:
      - campaigns
    patch:
      consumes:
      - application/json
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: string
      - description: Campaign name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCampaignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Campaign'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update campaign
      tags:
      - campaigns
  /api/entries:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Entry'
            type: array
      summary: List entries
      tags:
      - entries
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Entry'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create entry
      tags:
      - entries
  /api/entries/{id}:
    delete:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
      summary: Delete entry
      tags:
      - entries
    get:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Entry'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get entry
      tags:
      - entries
    patch:
      consumes:
      - application/json
      description: Partial update — send only the fields to change.
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (partial Entry)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.Entry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Entry'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update entry
      tags:
      - entries
  /api/entries/{id}/images:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Upload image attachment for an entry
      tags:
      - entries
  /api/entries/{id}/shares:
    get:
      description: Returns all users who have been granted access to an entry.
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.EntryShare'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List entry shares
      tags:
      - shares
    post:
      consumes:
      - application/json
      description: Grants access to an entry for one or more user IDs. Caller must
        be the entry owner.
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: '{ userIds: [\'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Share entry with users
      tags:
      - shares
  /api/entries/{id}/shares/{userId}:
    delete:
      description: Revokes a user's access to an entry. Caller must be the entry owner.
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Remove entry share
      tags:
      - shares
  /api/me/settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get the caller's settings
      tags:
      - settings
    patch:
      consumes:
      - application/json
      description: 'Shallow merge: each top-level key in the body overwrites the stored
        key; a null value deletes the key.'
      parameters:
      - description: Settings patch (JSON object)
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Merge settings into the caller's stored settings
      tags:
      - settings
  /api/member/entries/{id}/meta:
    put:
      consumes:
      - application/json
      description: Sets pinned status, sort order, and personal tags for an entry.
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: string
      - description: '{ pinned?, sortOrder?, tags? }'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MemberEntryMeta'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Upsert personal entry metadata
      tags:
      - member
  /api/member/folders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.MemberFolder'
            type: array
      summary: List my folders
      tags:
      - member
    post:
      consumes:
      - application/json
      parameters:
      - description: '{ name, color?, projectId? }'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.MemberFolder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a folder
      tags:
      - member
  /api/member/folders/{id}:
    delete:
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a folder
      tags:
      - member
    patch:
      consumes:
      - application/json
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: '{ name?, color?, sortOrder? }'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MemberFolder'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update a folder
      tags:
      - member
  /api/member/folders/{id}/entries:
    post:
      consumes:
      - application/json
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: '{ entryId }'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Add entry to folder
      tags:
      - member
  /api/member/folders/{id}/entries/{entryId}:
    delete:
      parameters:
      - description: Folder ID
        in: path
        name: id
        required: true
        type: string
      - description: Entry ID
        in: path
        name: entryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.OkResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Remove entry from folder
      tags:
      - member
...
swagger: "2.0"